requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
requests